test: test both ckb 2019 and ckb 2021 for ckb script features
