+++ script/src/verify.rs
@@ -43,6 +43,9 @@ use std::cell::RefCell;
 use std::collections::HashMap;
 use std::convert::TryFrom;
 
+#[cfg(test)]
+mod tests;
+
 pub enum ChunkState<'a> {
     Suspended(ResumableMachine<'a>),
     Completed(Cycle),
@@ -1009,1845 +1012,4 @@ fn wrapping_cycles_add(
 ) -> Result<Cycle, TransactionScriptError> {
     lhs.checked_add(rhs)
         .ok_or_else(|| ScriptError::CyclesOverflow(lhs, rhs).source(group))
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::type_id::TYPE_ID_CYCLES;
-    use byteorder::{ByteOrder, LittleEndian};
-    use ckb_crypto::secp::{Generator, Privkey, Pubkey, Signature};
-    use ckb_db::RocksDB;
-    use ckb_db_schema::COLUMNS;
-    use ckb_hash::{blake2b_256, new_blake2b};
-    use ckb_store::{data_loader_wrapper::DataLoaderWrapper, ChainDB};
-    use ckb_types::{
-        core::{
-            capacity_bytes, cell::CellMetaBuilder, hardfork::HardForkSwitch, Capacity, Cycle,
-            DepType, EpochNumber, EpochNumberWithFraction, HeaderView, ScriptHashType,
-            TransactionBuilder, TransactionInfo,
-        },
-        h256,
-        packed::{
-            Byte32, CellDep, CellInput, CellOutputBuilder, OutPoint, Script,
-            TransactionInfoBuilder, TransactionKeyBuilder, WitnessArgs,
-        },
-        H256,
-    };
-    use faster_hex::hex_encode;
-
-    use ckb_chain_spec::consensus::{
-        ConsensusBuilder, TWO_IN_TWO_OUT_BYTES, TWO_IN_TWO_OUT_CYCLES, TYPE_ID_CODE_HASH,
-    };
-    use ckb_error::assert_error_eq;
-    use ckb_test_chain_utils::{
-        always_success_cell, ckb_testnet_consensus, secp256k1_blake160_sighash_cell,
-        secp256k1_data_cell, type_lock_script_code_hash,
-    };
-    use std::fs::File;
-    use std::io::Read;
-    use std::path::Path;
-
-    const ALWAYS_SUCCESS_SCRIPT_CYCLE: u64 = 537;
-    const CYCLE_BOUND: Cycle = 200_000;
-
-    fn sha3_256<T: AsRef<[u8]>>(s: T) -> [u8; 32] {
-        use tiny_keccak::{Hasher, Sha3};
-        let mut output = [0; 32];
-        let mut sha3 = Sha3::v256();
-        sha3.update(s.as_ref());
-        sha3.finalize(&mut output);
-        output
-    }
-
-    fn open_cell_always_success() -> File {
-        File::open(Path::new(env!("CARGO_MANIFEST_DIR")).join("testdata/always_success")).unwrap()
-    }
-
-    fn open_cell_always_failure() -> File {
-        File::open(Path::new(env!("CARGO_MANIFEST_DIR")).join("testdata/always_failure")).unwrap()
-    }
-
-    fn new_store() -> ChainDB {
-        ChainDB::new(RocksDB::open_tmp(COLUMNS), Default::default())
-    }
-
-    fn random_keypair() -> (Privkey, Pubkey) {
-        Generator::random_keypair()
-    }
-
-    fn to_hex_pubkey(pubkey: &Pubkey) -> Vec<u8> {
-        let pubkey = pubkey.serialize();
-        let mut hex_pubkey = vec![0; pubkey.len() * 2];
-        hex_encode(&pubkey, &mut hex_pubkey).expect("hex pubkey");
-        hex_pubkey
-    }
-
-    fn to_hex_signature(signature: &Signature) -> Vec<u8> {
-        let signature_der = signature.serialize_der();
-        let mut hex_signature = vec![0; signature_der.len() * 2];
-        hex_encode(&signature_der, &mut hex_signature).expect("hex signature");
-        hex_signature
-    }
-
-    fn sign_args(args: &[u8], privkey: &Privkey) -> Signature {
-        let hash = sha3_256(sha3_256(args));
-        privkey.sign_recoverable(&hash.into()).unwrap()
-    }
-
-    fn default_transaction_info() -> TransactionInfo {
-        TransactionInfoBuilder::default()
-            .block_number(1u64.pack())
-            .block_epoch(0u64.pack())
-            .key(
-                TransactionKeyBuilder::default()
-                    .block_hash(Byte32::zero())
-                    .index(1u32.pack())
-                    .build(),
-            )
-            .build()
-            .unpack()
-    }
-
-    struct TransactionScriptsVerifierWithEnv {
-        store: ChainDB,
-        version_1_enabled_at: EpochNumber,
-        consensus: Consensus,
-    }
-
-    impl TransactionScriptsVerifierWithEnv {
-        fn new() -> Self {
-            let store = new_store();
-            let version_1_enabled_at = 10;
-            let hardfork_switch = HardForkSwitch::new_without_any_enabled()
-                .as_builder()
-                .rfc_pr_0234(version_1_enabled_at)
-                .rfc_pr_0237(version_1_enabled_at)
-                .build()
-                .unwrap();
-            let consensus = ConsensusBuilder::default()
-                .hardfork_switch(hardfork_switch)
-                .build();
-            Self {
-                store,
-                version_1_enabled_at,
-                consensus,
-            }
-        }
-
-        fn verify_without_limit(
-            &self,
-            version: ScriptVersion,
-            rtx: &ResolvedTransaction,
-        ) -> Result<Cycle, Error> {
-            self.verify(version, rtx, u64::MAX)
-        }
-
-        // If the max cycles is meaningless, please use `verify_without_limit`,
-        // so reviewers or developers can understand the intentions easier.
-        fn verify(
-            &self,
-            version: ScriptVersion,
-            rtx: &ResolvedTransaction,
-            max_cycles: Cycle,
-        ) -> Result<Cycle, Error> {
-            self.verify_map(version, rtx, |verifier| verifier.verify(max_cycles))
-        }
-
-        fn verify_map<R, F>(
-            &self,
-            version: ScriptVersion,
-            rtx: &ResolvedTransaction,
-            mut verify_func: F,
-        ) -> R
-        where
-            F: FnMut(TransactionScriptsVerifier<'_, DataLoaderWrapper<'_, ChainDB>>) -> R,
-        {
-            let data_loader = DataLoaderWrapper::new(&self.store);
-
-            let epoch = match version {
-                ScriptVersion::V0 => EpochNumberWithFraction::new(0, 0, 1),
-                ScriptVersion::V1 => EpochNumberWithFraction::new(self.version_1_enabled_at, 0, 1),
-            };
-            let header = HeaderView::new_advanced_builder()
-                .epoch(epoch.pack())
-                .build();
-            let tx_env = TxVerifyEnv::new_commit(&header);
-
-            let verifier =
-                TransactionScriptsVerifier::new(rtx, &self.consensus, &data_loader, &tx_env);
-            verify_func(verifier)
-        }
-    }
-
-    #[test]
-    fn check_always_success_hash() {
-        let (always_success_cell, always_success_cell_data, always_success_script) =
-            always_success_cell();
-        let output = CellOutputBuilder::default()
-            .capacity(capacity_bytes!(100).pack())
-            .lock(always_success_script.clone())
-            .build();
-        let input = CellInput::new(OutPoint::null(), 0);
-
-        let transaction = TransactionBuilder::default().input(input).build();
-
-        let dummy_cell = CellMetaBuilder::from_cell_output(output, Bytes::new())
-            .transaction_info(default_transaction_info())
-            .build();
-        let always_success_cell = CellMetaBuilder::from_cell_output(
-            always_success_cell.clone(),
-            always_success_cell_data.to_owned(),
-        )
-        .transaction_info(default_transaction_info())
-        .build();
-
-        let rtx = ResolvedTransaction {
-            transaction,
-            resolved_cell_deps: vec![always_success_cell],
-            resolved_inputs: vec![dummy_cell],
-            resolved_dep_groups: vec![],
-        };
-
-        let verifier = TransactionScriptsVerifierWithEnv::new();
-        let result = verifier.verify_without_limit(ScriptVersion::V0, &rtx);
-        assert!(result.is_ok());
-    }
-
-    #[test]
-    fn check_signature() {
-        let mut file = open_cell_always_success();
-        let mut buffer = Vec::new();
-        file.read_to_end(&mut buffer).unwrap();
-
-        let (privkey, pubkey) = random_keypair();
-        let mut args = b"foobar".to_vec();
-
-        let signature = sign_args(&args, &privkey);
-        args.extend(&to_hex_pubkey(&pubkey));
-        args.extend(&to_hex_signature(&signature));
-
-        let code_hash = blake2b_256(&buffer);
-        let dep_out_point = OutPoint::new(h256!("0x123").pack(), 8);
-        let cell_dep = CellDep::new_builder()
-            .out_point(dep_out_point.clone())
-            .build();
-        let data = Bytes::from(buffer);
-        let output = CellOutputBuilder::default()
-            .capacity(Capacity::bytes(data.len()).unwrap().pack())
-            .build();
-        let dep_cell = CellMetaBuilder::from_cell_output(output, data)
-            .transaction_info(default_transaction_info())
-            .out_point(dep_out_point)
-            .build();
-
-        let script = Script::new_builder()
-            .args(Bytes::from(args).pack())
-            .code_hash(code_hash.pack())
-            .hash_type(ScriptHashType::Data.into())
-            .build();
-        let input = CellInput::new(OutPoint::null(), 0);
-
-        let transaction = TransactionBuilder::default()
-            .input(input)
-            .cell_dep(cell_dep)
-            .build();
-
-        let output = CellOutputBuilder::default()
-            .capacity(capacity_bytes!(100).pack())
-            .lock(script)
-            .build();
-        let dummy_cell = CellMetaBuilder::from_cell_output(output, Bytes::new())
-            .transaction_info(default_transaction_info())
-            .build();
-
-        let rtx = ResolvedTransaction {
-            transaction,
-            resolved_cell_deps: vec![dep_cell],
-            resolved_inputs: vec![dummy_cell],
-            resolved_dep_groups: vec![],
-        };
-
-        let verifier = TransactionScriptsVerifierWithEnv::new();
-        let result = verifier.verify_without_limit(ScriptVersion::V0, &rtx);
-        assert!(result.is_ok());
-
-        // Not enough cycles
-        let max_cycles = ALWAYS_SUCCESS_SCRIPT_CYCLE - 1;
-        let result = verifier.verify(ScriptVersion::V0, &rtx, max_cycles);
-        assert_error_eq!(
-            result.unwrap_err(),
-            ScriptError::ExceededMaximumCycles(max_cycles).input_lock_script(0),
-        );
-
-        let result = verifier.verify_without_limit(ScriptVersion::V0, &rtx);
-        assert!(result.is_ok());
-    }
-
-    #[test]
-    fn check_signature_referenced_via_type_hash() {
-        let mut file = open_cell_always_success();
-        let mut buffer = Vec::new();
-        file.read_to_end(&mut buffer).unwrap();
-
-        let (privkey, pubkey) = random_keypair();
-        let mut args = b"foobar".to_vec();
-
-        let signature = sign_args(&args, &privkey);
-        args.extend(&to_hex_pubkey(&pubkey));
-        args.extend(&to_hex_signature(&signature));
-
-        let dep_out_point = OutPoint::new(h256!("0x123").pack(), 8);
-        let cell_dep = CellDep::new_builder()
-            .out_point(dep_out_point.clone())
-            .build();
-        let data = Bytes::from(buffer);
-        let output = CellOutputBuilder::default()
-            .capacity(Capacity::bytes(data.len()).unwrap().pack())
-            .type_(
-                Some(
-                    Script::new_builder()
-                        .code_hash(h256!("0x123456abcd90").pack())
-                        .hash_type(ScriptHashType::Data.into())
-                        .build(),
-                )
-                .pack(),
-            )
-            .build();
-        let type_hash = output.type_().to_opt().as_ref().unwrap().calc_script_hash();
-        let dep_cell = CellMetaBuilder::from_cell_output(output, data)
-            .transaction_info(default_transaction_info())
-            .out_point(dep_out_point)
-            .build();
-
-        let script = Script::new_builder()
-            .args(Bytes::from(args).pack())
-            .code_hash(type_hash)
-            .hash_type(ScriptHashType::Type.into())
-            .build();
-        let input = CellInput::new(OutPoint::null(), 0);
-
-        let transaction = TransactionBuilder::default()
-            .input(input)
-            .cell_dep(cell_dep)
-            .build();
-
-        let output = CellOutputBuilder::default()
-            .capacity(capacity_bytes!(100).pack())
-            .lock(script)
-            .build();
-        let dummy_cell = CellMetaBuilder::from_cell_output(output, Bytes::new())
-            .transaction_info(default_transaction_info())
-            .build();
-
-        let rtx = ResolvedTransaction {
-            transaction,
-            resolved_cell_deps: vec![dep_cell],
-            resolved_inputs: vec![dummy_cell],
-            resolved_dep_groups: vec![],
-        };
-
-        let verifier = TransactionScriptsVerifierWithEnv::new();
-        let result = verifier.verify_without_limit(ScriptVersion::V0, &rtx);
-        assert!(result.is_ok());
-    }
-
-    #[test]
-    fn check_signature_referenced_via_type_hash_failure_with_multiple_matches() {
-        let mut file = open_cell_always_success();
-        let mut buffer = Vec::new();
-        file.read_to_end(&mut buffer).unwrap();
-        let data = Bytes::from(buffer);
-
-        let (privkey, pubkey) = random_keypair();
-        let mut args = b"foobar".to_vec();
-
-        let signature = sign_args(&args, &privkey);
-        args.extend(&to_hex_pubkey(&pubkey));
-        args.extend(&to_hex_signature(&signature));
-
-        let dep_out_point = OutPoint::new(h256!("0x123").pack(), 8);
-        let cell_dep = CellDep::new_builder()
-            .out_point(dep_out_point.clone())
-            .build();
-        let output = CellOutputBuilder::default()
-            .capacity(Capacity::bytes(data.len()).unwrap().pack())
-            .type_(
-                Some(
-                    Script::new_builder()
-                        .code_hash(h256!("0x123456abcd90").pack())
-                        .hash_type(ScriptHashType::Data.into())
-                        .build(),
-                )
-                .pack(),
-            )
-            .build();
-        let type_hash = output.type_().to_opt().as_ref().unwrap().calc_script_hash();
-        let dep_cell = CellMetaBuilder::from_cell_output(output, data.clone())
-            .transaction_info(default_transaction_info())
-            .out_point(dep_out_point)
-            .build();
-
-        let dep_out_point2 = OutPoint::new(h256!("0x1234").pack(), 8);
-        let cell_dep2 = CellDep::new_builder()
-            .out_point(dep_out_point2.clone())
-            .build();
-        let output2 = CellOutputBuilder::default()
-            .capacity(Capacity::bytes(data.len()).unwrap().pack())
-            .type_(
-                Some(
-                    Script::new_builder()
-                        .code_hash(h256!("0x123456abcd90").pack())
-                        .hash_type(ScriptHashType::Data.into())
-                        .build(),
-                )
-                .pack(),
-            )
-            .build();
-        let dep_cell2 = CellMetaBuilder::from_cell_output(output2, data)
-            .transaction_info(default_transaction_info())
-            .out_point(dep_out_point2)
-            .build();
-
-        let script = Script::new_builder()
-            .args(Bytes::from(args).pack())
-            .code_hash(type_hash)
-            .hash_type(ScriptHashType::Type.into())
-            .build();
-        let input = CellInput::new(OutPoint::null(), 0);
-
-        let transaction = TransactionBuilder::default()
-            .input(input)
-            .cell_dep(cell_dep)
-            .cell_dep(cell_dep2)
-            .build();
-
-        let output = CellOutputBuilder::default()
-            .capacity(capacity_bytes!(100).pack())
-            .lock(script)
-            .build();
-        let dummy_cell = CellMetaBuilder::from_cell_output(output, Bytes::new())
-            .transaction_info(default_transaction_info())
-            .build();
-
-        let rtx = ResolvedTransaction {
-            transaction,
-            resolved_cell_deps: vec![dep_cell, dep_cell2],
-            resolved_inputs: vec![dummy_cell],
-            resolved_dep_groups: vec![],
-        };
-
-        let verifier = TransactionScriptsVerifierWithEnv::new();
-        let result = verifier.verify_without_limit(ScriptVersion::V0, &rtx);
-        assert_error_eq!(
-            result.unwrap_err(),
-            ScriptError::MultipleMatches.input_lock_script(0),
-        );
-    }
-
-    #[test]
-    fn check_invalid_signature() {
-        let mut file = open_cell_always_failure();
-        let mut buffer = Vec::new();
-        file.read_to_end(&mut buffer).unwrap();
-
-        let (privkey, pubkey) = random_keypair();
-        let mut args = b"foobar".to_vec();
-
-        let signature = sign_args(&args, &privkey);
-
-        // This line makes the verification invalid
-        args.extend(&b"extrastring".to_vec());
-        args.extend(&to_hex_pubkey(&pubkey));
-        args.extend(&to_hex_signature(&signature));
-
-        let code_hash = blake2b_256(&buffer);
-        let dep_out_point = OutPoint::new(h256!("0x123").pack(), 8);
-        let cell_dep = CellDep::new_builder().out_point(dep_out_point).build();
-        let data = Bytes::from(buffer);
-        let output = CellOutputBuilder::default()
-            .capacity(Capacity::bytes(data.len()).unwrap().pack())
-            .build();
-        let dep_cell = CellMetaBuilder::from_cell_output(output, data)
-            .transaction_info(default_transaction_info())
-            .build();
-
-        let script = Script::new_builder()
-            .args(Bytes::from(args).pack())
-            .code_hash(code_hash.pack())
-            .hash_type(ScriptHashType::Data.into())
-            .build();
-        let input = CellInput::new(OutPoint::null(), 0);
-
-        let transaction = TransactionBuilder::default()
-            .input(input)
-            .cell_dep(cell_dep)
-            .build();
-
-        let output = CellOutputBuilder::default()
-            .capacity(capacity_bytes!(100).pack())
-            .lock(script.clone())
-            .build();
-        let dummy_cell = CellMetaBuilder::from_cell_output(output, Bytes::new())
-            .transaction_info(default_transaction_info())
-            .build();
-
-        let rtx = ResolvedTransaction {
-            transaction,
-            resolved_cell_deps: vec![dep_cell],
-            resolved_inputs: vec![dummy_cell],
-            resolved_dep_groups: vec![],
-        };
-
-        let verifier = TransactionScriptsVerifierWithEnv::new();
-        let result = verifier.verify_without_limit(ScriptVersion::V0, &rtx);
-        assert_error_eq!(
-            result.unwrap_err(),
-            ScriptError::validation_failure(&script, -1).input_lock_script(0),
-        );
-    }
-
-    #[test]
-    fn check_invalid_dep_reference() {
-        let mut file = open_cell_always_success();
-        let mut buffer = Vec::new();
-        file.read_to_end(&mut buffer).unwrap();
-
-        let (privkey, pubkey) = random_keypair();
-        let mut args = b"foobar".to_vec();
-        let signature = sign_args(&args, &privkey);
-        args.extend(&to_hex_pubkey(&pubkey));
-        args.extend(&to_hex_signature(&signature));
-
-        let dep_out_point = OutPoint::new(h256!("0x123").pack(), 8);
-        let cell_dep = CellDep::new_builder().out_point(dep_out_point).build();
-
-        let script = Script::new_builder()
-            .args(Bytes::from(args).pack())
-            .code_hash(blake2b_256(&buffer).pack())
-            .hash_type(ScriptHashType::Data.into())
-            .build();
-        let input = CellInput::new(OutPoint::null(), 0);
-
-        let transaction = TransactionBuilder::default()
-            .input(input)
-            .cell_dep(cell_dep)
-            .build();
-
-        let output = CellOutputBuilder::default()
-            .capacity(capacity_bytes!(100).pack())
-            .lock(script)
-            .build();
-        let dummy_cell = CellMetaBuilder::from_cell_output(output, Bytes::new())
-            .transaction_info(default_transaction_info())
-            .build();
-
-        let rtx = ResolvedTransaction {
-            transaction,
-            resolved_cell_deps: vec![],
-            resolved_inputs: vec![dummy_cell],
-            resolved_dep_groups: vec![],
-        };
-
-        let verifier = TransactionScriptsVerifierWithEnv::new();
-        let result = verifier.verify_without_limit(ScriptVersion::V0, &rtx);
-        assert_error_eq!(
-            result.unwrap_err(),
-            ScriptError::InvalidCodeHash.input_lock_script(0),
-        );
-    }
-
-    #[test]
-    fn check_output_contract() {
-        let mut file = open_cell_always_success();
-        let mut buffer = Vec::new();
-        file.read_to_end(&mut buffer).unwrap();
-
-        let (privkey, pubkey) = random_keypair();
-        let mut args = b"foobar".to_vec();
-        let signature = sign_args(&args, &privkey);
-        args.extend(&to_hex_pubkey(&pubkey));
-        args.extend(&to_hex_signature(&signature));
-
-        let input = CellInput::new(OutPoint::null(), 0);
-        let (always_success_cell, always_success_cell_data, always_success_script) =
-            always_success_cell();
-        let output = CellOutputBuilder::default()
-            .capacity(capacity_bytes!(100).pack())
-            .lock(always_success_script.clone())
-            .build();
-        let dummy_cell = CellMetaBuilder::from_cell_output(output, Bytes::new())
-            .transaction_info(default_transaction_info())
-            .build();
-        let always_success_cell = CellMetaBuilder::from_cell_output(
-            always_success_cell.clone(),
-            always_success_cell_data.to_owned(),
-        )
-        .transaction_info(default_transaction_info())
-        .build();
-
-        let script = Script::new_builder()
-            .args(Bytes::from(args).pack())
-            .code_hash(blake2b_256(&buffer).pack())
-            .hash_type(ScriptHashType::Data.into())
-            .build();
-        let output_data = Bytes::default();
-        let output = CellOutputBuilder::default()
-            .lock(
-                Script::new_builder()
-                    .hash_type(ScriptHashType::Data.into())
-                    .build(),
-            )
-            .type_(Some(script).pack())
-            .build();
-
-        let dep_out_point = OutPoint::new(h256!("0x123").pack(), 8);
-        let cell_dep = CellDep::new_builder()
-            .out_point(dep_out_point.clone())
-            .build();
-        let dep_cell = {
-            let data = Bytes::from(buffer);
-            let output = CellOutputBuilder::default()
-                .capacity(Capacity::bytes(data.len()).unwrap().pack())
-                .build();
-            CellMetaBuilder::from_cell_output(output, data)
-                .transaction_info(default_transaction_info())
-                .out_point(dep_out_point)
-                .build()
-        };
-
-        let transaction = TransactionBuilder::default()
-            .input(input)
-            .output(output)
-            .output_data(output_data.pack())
-            .cell_dep(cell_dep)
-            .build();
-
-        let rtx = ResolvedTransaction {
-            transaction,
-            resolved_cell_deps: vec![dep_cell, always_success_cell],
-            resolved_inputs: vec![dummy_cell],
-            resolved_dep_groups: vec![],
-        };
-
-        let verifier = TransactionScriptsVerifierWithEnv::new();
-        let result = verifier.verify_without_limit(ScriptVersion::V0, &rtx);
-        assert!(result.is_ok());
-    }
-
-    #[test]
-    fn check_invalid_output_contract() {
-        let mut file = open_cell_always_failure();
-        let mut buffer = Vec::new();
-        file.read_to_end(&mut buffer).unwrap();
-
-        let (privkey, pubkey) = random_keypair();
-        let mut args = b"foobar".to_vec();
-
-        let signature = sign_args(&args, &privkey);
-        // This line makes the verification invalid
-        args.extend(&b"extrastring".to_vec());
-        args.extend(&to_hex_pubkey(&pubkey));
-        args.extend(&to_hex_signature(&signature));
-
-        let input = CellInput::new(OutPoint::null(), 0);
-        let (always_success_cell, always_success_cell_data, always_success_script) =
-            always_success_cell();
-        let output = CellOutputBuilder::default()
-            .capacity(capacity_bytes!(100).pack())
-            .lock(always_success_script.clone())
-            .build();
-        let dummy_cell = CellMetaBuilder::from_cell_output(output, Bytes::new())
-            .transaction_info(default_transaction_info())
-            .build();
-        let always_success_cell = CellMetaBuilder::from_cell_output(
-            always_success_cell.to_owned(),
-            always_success_cell_data.to_owned(),
-        )
-        .transaction_info(default_transaction_info())
-        .build();
-
-        let script = Script::new_builder()
-            .args(Bytes::from(args).pack())
-            .code_hash(blake2b_256(&buffer).pack())
-            .hash_type(ScriptHashType::Data.into())
-            .build();
-        let output = CellOutputBuilder::default()
-            .type_(Some(script.clone()).pack())
-            .build();
-
-        let dep_out_point = OutPoint::new(h256!("0x123").pack(), 8);
-        let cell_dep = CellDep::new_builder().out_point(dep_out_point).build();
-        let dep_cell = {
-            let dep_cell_data = Bytes::from(buffer);
-            let output = CellOutputBuilder::default()
-                .capacity(Capacity::bytes(dep_cell_data.len()).unwrap().pack())
-                .build();
-            CellMetaBuilder::from_cell_output(output, dep_cell_data)
-                .transaction_info(default_transaction_info())
-                .build()
-        };
-
-        let transaction = TransactionBuilder::default()
-            .input(input)
-            .output(output)
-            .output_data(Bytes::new().pack())
-            .cell_dep(cell_dep)
-            .build();
-
-        let rtx = ResolvedTransaction {
-            transaction,
-            resolved_cell_deps: vec![dep_cell, always_success_cell],
-            resolved_inputs: vec![dummy_cell],
-            resolved_dep_groups: vec![],
-        };
-
-        let verifier = TransactionScriptsVerifierWithEnv::new();
-        let result = verifier.verify_without_limit(ScriptVersion::V0, &rtx);
-        assert_error_eq!(
-            result.unwrap_err(),
-            ScriptError::validation_failure(&script, -1).output_type_script(0),
-        );
-    }
-
-    #[test]
-    fn check_same_lock_and_type_script_are_executed_twice() {
-        let mut file = open_cell_always_success();
-        let mut buffer = Vec::new();
-        file.read_to_end(&mut buffer).unwrap();
-
-        let privkey = Privkey::from_slice(&[1; 32][..]);
-        let pubkey = privkey.pubkey().unwrap();
-        let mut args = b"foobar".to_vec();
-
-        let signature = sign_args(&args, &privkey);
-        args.extend(&to_hex_pubkey(&pubkey));
-        args.extend(&to_hex_signature(&signature));
-
-        let script = Script::new_builder()
-            .args(Bytes::from(args).pack())
-            .code_hash(blake2b_256(&buffer).pack())
-            .hash_type(ScriptHashType::Data.into())
-            .build();
-
-        let dep_out_point = OutPoint::new(h256!("0x123").pack(), 8);
-        let cell_dep = CellDep::new_builder()
-            .out_point(dep_out_point.clone())
-            .build();
-        let data = Bytes::from(buffer);
-        let output = CellOutputBuilder::default()
-            .capacity(Capacity::bytes(data.len()).unwrap().pack())
-            .build();
-        let dep_cell = CellMetaBuilder::from_cell_output(output, data)
-            .transaction_info(default_transaction_info())
-            .out_point(dep_out_point)
-            .build();
-
-        let transaction = TransactionBuilder::default()
-            .input(CellInput::new(OutPoint::null(), 0))
-            .cell_dep(cell_dep)
-            .build();
-
-        // The lock and type scripts here are both executed.
-        let output = CellOutputBuilder::default()
-            .capacity(capacity_bytes!(100).pack())
-            .lock(script.clone())
-            .type_(Some(script).pack())
-            .build();
-        let dummy_cell = CellMetaBuilder::from_cell_output(output, Bytes::new())
-            .transaction_info(default_transaction_info())
-            .build();
-
-        let rtx = ResolvedTransaction {
-            transaction,
-            resolved_cell_deps: vec![dep_cell],
-            resolved_inputs: vec![dummy_cell],
-            resolved_dep_groups: vec![],
-        };
-
-        let verifier = TransactionScriptsVerifierWithEnv::new();
-        // Cycles can tell that both lock and type scripts are executed
-        let result = verifier.verify_without_limit(ScriptVersion::V0, &rtx);
-        assert_eq!(result.ok(), Some(ALWAYS_SUCCESS_SCRIPT_CYCLE * 2));
-    }
-
-    #[test]
-    fn check_type_id_one_in_one_out() {
-        let (always_success_cell, always_success_cell_data, always_success_script) =
-            always_success_cell();
-        let always_success_out_point = OutPoint::new(h256!("0x11").pack(), 0);
-
-        let type_id_script = Script::new_builder()
-            .args(Bytes::from(h256!("0x1111").as_ref()).pack())
-            .code_hash(TYPE_ID_CODE_HASH.pack())
-            .hash_type(ScriptHashType::Type.into())
-            .build();
-
-        let input = CellInput::new(OutPoint::new(h256!("0x1234").pack(), 8), 0);
-        let input_cell = CellOutputBuilder::default()
-            .capacity(capacity_bytes!(1000).pack())
-            .lock(always_success_script.clone())
-            .type_(Some(type_id_script.clone()).pack())
-            .build();
-
-        let output_cell = CellOutputBuilder::default()
-            .capacity(capacity_bytes!(990).pack())
-            .lock(always_success_script.clone())
-            .type_(Some(type_id_script).pack())
-            .build();
-
-        let transaction = TransactionBuilder::default()
-            .input(input.clone())
-            .output(output_cell)
-            .cell_dep(
-                CellDep::new_builder()
-                    .out_point(always_success_out_point.clone())
-                    .build(),
-            )
-            .build();
-
-        let resolved_input_cell = CellMetaBuilder::from_cell_output(input_cell, Bytes::new())
-            .out_point(input.previous_output())
-            .build();
-        let resolved_always_success_cell = CellMetaBuilder::from_cell_output(
-            always_success_cell.clone(),
-            always_success_cell_data.to_owned(),
-        )
-        .out_point(always_success_out_point)
-        .build();
-
-        let rtx = ResolvedTransaction {
-            transaction,
-            resolved_cell_deps: vec![resolved_always_success_cell],
-            resolved_inputs: vec![resolved_input_cell],
-            resolved_dep_groups: vec![],
-        };
-
-        let max_cycles = TYPE_ID_CYCLES * 2;
-        let verifier = TransactionScriptsVerifierWithEnv::new();
-        let result = verifier.verify(ScriptVersion::V0, &rtx, max_cycles);
-        assert!(
-            result.is_ok(),
-            "expect ok, but got {:?}",
-            result.unwrap_err()
-        );
-    }
-
-    #[test]
-    fn check_type_id_one_in_one_out_not_enough_cycles() {
-        let (always_success_cell, always_success_cell_data, always_success_script) =
-            always_success_cell();
-        let always_success_out_point = OutPoint::new(h256!("0x11").pack(), 0);
-
-        let type_id_script = Script::new_builder()
-            .args(Bytes::from(h256!("0x1111").as_ref()).pack())
-            .code_hash(TYPE_ID_CODE_HASH.pack())
-            .hash_type(ScriptHashType::Type.into())
-            .build();
-
-        let input = CellInput::new(OutPoint::new(h256!("0x1234").pack(), 8), 0);
-        let input_cell = CellOutputBuilder::default()
-            .capacity(capacity_bytes!(1000).pack())
-            .lock(always_success_script.clone())
-            .type_(Some(type_id_script.clone()).pack())
-            .build();
-
-        let output_cell = CellOutputBuilder::default()
-            .capacity(capacity_bytes!(990).pack())
-            .lock(always_success_script.clone())
-            .type_(Some(type_id_script).pack())
-            .build();
-
-        let transaction = TransactionBuilder::default()
-            .input(input.clone())
-            .output(output_cell)
-            .cell_dep(
-                CellDep::new_builder()
-                    .out_point(always_success_out_point.clone())
-                    .build(),
-            )
-            .build();
-
-        let resolved_input_cell = CellMetaBuilder::from_cell_output(input_cell, Bytes::new())
-            .out_point(input.previous_output())
-            .build();
-        let resolved_always_success_cell = CellMetaBuilder::from_cell_output(
-            always_success_cell.clone(),
-            always_success_cell_data.to_owned(),
-        )
-        .out_point(always_success_out_point)
-        .build();
-
-        let rtx = ResolvedTransaction {
-            transaction,
-            resolved_cell_deps: vec![resolved_always_success_cell],
-            resolved_inputs: vec![resolved_input_cell],
-            resolved_dep_groups: vec![],
-        };
-
-        let max_cycles = TYPE_ID_CYCLES - 1;
-        let verifier = TransactionScriptsVerifierWithEnv::new();
-        // two groups need exec, so cycles not TYPE_ID_CYCLES - 1
-        let result = verifier.verify(ScriptVersion::V0, &rtx, max_cycles);
-        assert_error_eq!(
-            result.unwrap_err(),
-            ScriptError::ExceededMaximumCycles(TYPE_ID_CYCLES - ALWAYS_SUCCESS_SCRIPT_CYCLE - 1)
-                .input_type_script(0),
-        );
-    }
-
-    #[test]
-    fn check_type_id_creation() {
-        let (always_success_cell, always_success_cell_data, always_success_script) =
-            always_success_cell();
-        let always_success_out_point = OutPoint::new(h256!("0x11").pack(), 0);
-
-        let input = CellInput::new(OutPoint::new(h256!("0x1234").pack(), 8), 0);
-        let input_cell = CellOutputBuilder::default()
-            .capacity(capacity_bytes!(1000).pack())
-            .lock(always_success_script.clone())
-            .build();
-
-        let input_hash = {
-            let mut blake2b = new_blake2b();
-            blake2b.update(input.as_slice());
-            blake2b.update(&0u64.to_le_bytes());
-            let mut ret = [0; 32];
-            blake2b.finalize(&mut ret);
-            Bytes::from(ret.to_vec())
-        };
-
-        let type_id_script = Script::new_builder()
-            .args(input_hash.pack())
-            .code_hash(TYPE_ID_CODE_HASH.pack())
-            .hash_type(ScriptHashType::Type.into())
-            .build();
-
-        let output_cell = CellOutputBuilder::default()
-            .capacity(capacity_bytes!(990).pack())
-            .lock(always_success_script.clone())
-            .type_(Some(type_id_script).pack())
-            .build();
-
-        let transaction = TransactionBuilder::default()
-            .input(input.clone())
-            .output(output_cell)
-            .cell_dep(
-                CellDep::new_builder()
-                    .out_point(always_success_out_point.clone())
-                    .build(),
-            )
-            .build();
-
-        let resolved_input_cell = CellMetaBuilder::from_cell_output(input_cell, Bytes::new())
-            .out_point(input.previous_output())
-            .build();
-        let resolved_always_success_cell = CellMetaBuilder::from_cell_output(
-            always_success_cell.clone(),
-            always_success_cell_data.to_owned(),
-        )
-        .out_point(always_success_out_point)
-        .build();
-
-        let rtx = ResolvedTransaction {
-            transaction,
-            resolved_cell_deps: vec![resolved_always_success_cell],
-            resolved_inputs: vec![resolved_input_cell],
-            resolved_dep_groups: vec![],
-        };
-
-        let verifier = TransactionScriptsVerifierWithEnv::new();
-        let result = verifier.verify_without_limit(ScriptVersion::V0, &rtx);
-        assert!(result.is_ok());
-    }
-
-    #[test]
-    fn check_type_id_termination() {
-        let (always_success_cell, always_success_cell_data, always_success_script) =
-            always_success_cell();
-        let always_success_out_point = OutPoint::new(h256!("0x11").pack(), 0);
-
-        let type_id_script = Script::new_builder()
-            .args(Bytes::from(h256!("0x1111").as_ref()).pack())
-            .code_hash(TYPE_ID_CODE_HASH.pack())
-            .hash_type(ScriptHashType::Type.into())
-            .build();
-
-        let input = CellInput::new(OutPoint::new(h256!("0x1234").pack(), 8), 0);
-        let input_cell = CellOutputBuilder::default()
-            .capacity(capacity_bytes!(1000).pack())
-            .lock(always_success_script.clone())
-            .type_(Some(type_id_script).pack())
-            .build();
-
-        let output_cell = CellOutputBuilder::default()
-            .capacity(capacity_bytes!(990).pack())
-            .lock(always_success_script.clone())
-            .build();
-
-        let transaction = TransactionBuilder::default()
-            .input(input.clone())
-            .output(output_cell)
-            .cell_dep(
-                CellDep::new_builder()
-                    .out_point(always_success_out_point.clone())
-                    .build(),
-            )
-            .build();
-
-        let resolved_input_cell = CellMetaBuilder::from_cell_output(input_cell, Bytes::new())
-            .out_point(input.previous_output())
-            .build();
-        let resolved_always_success_cell = CellMetaBuilder::from_cell_output(
-            always_success_cell.clone(),
-            always_success_cell_data.to_owned(),
-        )
-        .out_point(always_success_out_point)
-        .build();
-
-        let rtx = ResolvedTransaction {
-            transaction,
-            resolved_cell_deps: vec![resolved_always_success_cell],
-            resolved_inputs: vec![resolved_input_cell],
-            resolved_dep_groups: vec![],
-        };
-
-        let verifier = TransactionScriptsVerifierWithEnv::new();
-        let result = verifier.verify_without_limit(ScriptVersion::V0, &rtx);
-        assert!(result.is_ok());
-    }
-
-    #[test]
-    fn check_type_id_invalid_creation() {
-        let (always_success_cell, always_success_cell_data, always_success_script) =
-            always_success_cell();
-        let always_success_out_point = OutPoint::new(h256!("0x11").pack(), 0);
-
-        let input = CellInput::new(OutPoint::new(h256!("0x1234").pack(), 8), 0);
-        let input_cell = CellOutputBuilder::default()
-            .capacity(capacity_bytes!(1000).pack())
-            .lock(always_success_script.clone())
-            .build();
-
-        let input_hash = {
-            let mut blake2b = new_blake2b();
-            blake2b.update(&input.previous_output().tx_hash().as_bytes());
-            let mut buf = [0; 4];
-            LittleEndian::write_u32(&mut buf, input.previous_output().index().unpack());
-            blake2b.update(&buf[..]);
-            let mut buf = [0; 8];
-            LittleEndian::write_u64(&mut buf, 0);
-            blake2b.update(&buf[..]);
-            blake2b.update(b"unnecessary data");
-            let mut ret = [0; 32];
-            blake2b.finalize(&mut ret);
-            Bytes::from(ret.to_vec())
-        };
-
-        let type_id_script = Script::new_builder()
-            .args(input_hash.pack())
-            .code_hash(TYPE_ID_CODE_HASH.pack())
-            .hash_type(ScriptHashType::Type.into())
-            .build();
-
-        let output_cell = CellOutputBuilder::default()
-            .capacity(capacity_bytes!(990).pack())
-            .lock(always_success_script.clone())
-            .type_(Some(type_id_script.clone()).pack())
-            .build();
-
-        let transaction = TransactionBuilder::default()
-            .input(input.clone())
-            .output(output_cell)
-            .cell_dep(
-                CellDep::new_builder()
-                    .out_point(always_success_out_point.clone())
-                    .build(),
-            )
-            .build();
-
-        let resolved_input_cell = CellMetaBuilder::from_cell_output(input_cell, Bytes::new())
-            .out_point(input.previous_output())
-            .build();
-        let resolved_always_success_cell = CellMetaBuilder::from_cell_output(
-            always_success_cell.clone(),
-            always_success_cell_data.to_owned(),
-        )
-        .out_point(always_success_out_point)
-        .build();
-
-        let rtx = ResolvedTransaction {
-            transaction,
-            resolved_cell_deps: vec![resolved_always_success_cell],
-            resolved_inputs: vec![resolved_input_cell],
-            resolved_dep_groups: vec![],
-        };
-
-        let verifier = TransactionScriptsVerifierWithEnv::new();
-        let result = verifier.verify_without_limit(ScriptVersion::V0, &rtx);
-        assert_error_eq!(
-            result.unwrap_err(),
-            ScriptError::validation_failure(&type_id_script, -3).output_type_script(0),
-        );
-    }
-
-    #[test]
-    fn check_type_id_invalid_creation_length() {
-        let (always_success_cell, always_success_cell_data, always_success_script) =
-            always_success_cell();
-        let always_success_out_point = OutPoint::new(h256!("0x11").pack(), 0);
-
-        let input = CellInput::new(OutPoint::new(h256!("0x1234").pack(), 8), 0);
-        let input_cell = CellOutputBuilder::default()
-            .capacity(capacity_bytes!(1000).pack())
-            .lock(always_success_script.clone())
-            .build();
-
-        let input_hash = {
-            let mut blake2b = new_blake2b();
-            blake2b.update(&input.previous_output().tx_hash().as_bytes());
-            let mut buf = [0; 4];
-            LittleEndian::write_u32(&mut buf, input.previous_output().index().unpack());
-            blake2b.update(&buf[..]);
-            let mut buf = [0; 8];
-            LittleEndian::write_u64(&mut buf, 0);
-            blake2b.update(&buf[..]);
-            let mut ret = [0; 32];
-            blake2b.finalize(&mut ret);
-
-            let mut buf = vec![];
-            buf.extend_from_slice(&ret[..]);
-            buf.extend_from_slice(b"abc");
-            Bytes::from(buf)
-        };
-
-        let type_id_script = Script::new_builder()
-            .args(input_hash.pack())
-            .code_hash(TYPE_ID_CODE_HASH.pack())
-            .hash_type(ScriptHashType::Type.into())
-            .build();
-
-        let output_cell = CellOutputBuilder::default()
-            .capacity(capacity_bytes!(990).pack())
-            .lock(always_success_script.clone())
-            .type_(Some(type_id_script.clone()).pack())
-            .build();
-
-        let transaction = TransactionBuilder::default()
-            .input(input.clone())
-            .output(output_cell)
-            .cell_dep(
-                CellDep::new_builder()
-                    .out_point(always_success_out_point.clone())
-                    .build(),
-            )
-            .build();
-
-        let resolved_input_cell = CellMetaBuilder::from_cell_output(input_cell, Bytes::new())
-            .out_point(input.previous_output())
-            .build();
-        let resolved_always_success_cell = CellMetaBuilder::from_cell_output(
-            always_success_cell.clone(),
-            always_success_cell_data.to_owned(),
-        )
-        .out_point(always_success_out_point)
-        .build();
-
-        let rtx = ResolvedTransaction {
-            transaction,
-            resolved_cell_deps: vec![resolved_always_success_cell],
-            resolved_inputs: vec![resolved_input_cell],
-            resolved_dep_groups: vec![],
-        };
-
-        let verifier = TransactionScriptsVerifierWithEnv::new();
-        let result = verifier.verify_without_limit(ScriptVersion::V0, &rtx);
-        assert_error_eq!(
-            result.unwrap_err(),
-            ScriptError::validation_failure(&type_id_script, -1).output_type_script(0),
-        );
-    }
-
-    #[test]
-    fn check_type_id_one_in_two_out() {
-        let (always_success_cell, always_success_cell_data, always_success_script) =
-            always_success_cell();
-        let always_success_out_point = OutPoint::new(h256!("0x11").pack(), 0);
-
-        let type_id_script = Script::new_builder()
-            .args(Bytes::from(h256!("0x1111").as_ref()).pack())
-            .code_hash(TYPE_ID_CODE_HASH.pack())
-            .hash_type(ScriptHashType::Type.into())
-            .build();
-
-        let input = CellInput::new(OutPoint::new(h256!("0x1234").pack(), 8), 0);
-        let input_cell = CellOutputBuilder::default()
-            .capacity(capacity_bytes!(2000).pack())
-            .lock(always_success_script.clone())
-            .type_(Some(type_id_script.clone()).pack())
-            .build();
-
-        let output_cell = CellOutputBuilder::default()
-            .capacity(capacity_bytes!(990).pack())
-            .lock(always_success_script.clone())
-            .type_(Some(type_id_script.clone()).pack())
-            .build();
-        let output_cell2 = CellOutputBuilder::default()
-            .capacity(capacity_bytes!(990).pack())
-            .lock(always_success_script.clone())
-            .type_(Some(type_id_script.clone()).pack())
-            .build();
-
-        let transaction = TransactionBuilder::default()
-            .input(input.clone())
-            .output(output_cell)
-            .output(output_cell2)
-            .cell_dep(
-                CellDep::new_builder()
-                    .out_point(always_success_out_point.clone())
-                    .build(),
-            )
-            .build();
-
-        let resolved_input_cell = CellMetaBuilder::from_cell_output(input_cell, Bytes::new())
-            .out_point(input.previous_output())
-            .build();
-        let resolved_always_success_cell = CellMetaBuilder::from_cell_output(
-            always_success_cell.clone(),
-            always_success_cell_data.to_owned(),
-        )
-        .out_point(always_success_out_point)
-        .build();
-
-        let rtx = ResolvedTransaction {
-            transaction,
-            resolved_cell_deps: vec![resolved_always_success_cell],
-            resolved_inputs: vec![resolved_input_cell],
-            resolved_dep_groups: vec![],
-        };
-
-        let max_cycles = TYPE_ID_CYCLES * 2;
-        let verifier = TransactionScriptsVerifierWithEnv::new();
-        let result = verifier.verify(ScriptVersion::V0, &rtx, max_cycles);
-        assert_error_eq!(
-            result.unwrap_err(),
-            ScriptError::validation_failure(&type_id_script, -2).input_type_script(0),
-        );
-    }
-
-    #[test]
-    fn check_typical_secp256k1_blake160_2_in_2_out_tx() {
-        let consensus = ckb_testnet_consensus();
-        let dep_group_tx_hash = consensus.genesis_block().transactions()[1].hash();
-        let secp_out_point = OutPoint::new(dep_group_tx_hash, 0);
-
-        let cell_dep = CellDep::new_builder()
-            .out_point(secp_out_point)
-            .dep_type(DepType::DepGroup.into())
-            .build();
-
-        let input1 = CellInput::new(OutPoint::new(h256!("0x1234").pack(), 0), 0);
-        let input2 = CellInput::new(OutPoint::new(h256!("0x1111").pack(), 0), 0);
-
-        let mut generator = Generator::non_crypto_safe_prng(42);
-        let privkey = generator.gen_privkey();
-        let pubkey_data = privkey.pubkey().expect("Get pubkey failed").serialize();
-        let lock_arg = Bytes::from((&blake2b_256(&pubkey_data)[0..20]).to_owned());
-        let privkey2 = generator.gen_privkey();
-        let pubkey_data2 = privkey2.pubkey().expect("Get pubkey failed").serialize();
-        let lock_arg2 = Bytes::from((&blake2b_256(&pubkey_data2)[0..20]).to_owned());
-
-        let lock = Script::new_builder()
-            .args(lock_arg.pack())
-            .code_hash(type_lock_script_code_hash().pack())
-            .hash_type(ScriptHashType::Type.into())
-            .build();
-
-        let lock2 = Script::new_builder()
-            .args(lock_arg2.pack())
-            .code_hash(type_lock_script_code_hash().pack())
-            .hash_type(ScriptHashType::Type.into())
-            .build();
-
-        let output1 = CellOutput::new_builder()
-            .capacity(capacity_bytes!(100).pack())
-            .lock(lock.clone())
-            .build();
-        let output2 = CellOutput::new_builder()
-            .capacity(capacity_bytes!(100).pack())
-            .lock(lock2.clone())
-            .build();
-        let tx = TransactionBuilder::default()
-            .cell_dep(cell_dep)
-            .input(input1.clone())
-            .input(input2.clone())
-            .output(output1)
-            .output(output2)
-            .output_data(Default::default())
-            .output_data(Default::default())
-            .build();
-
-        let tx_hash: H256 = tx.hash().unpack();
-        // sign input1
-        let witness = {
-            WitnessArgs::new_builder()
-                .lock(Some(Bytes::from(vec![0u8; 65])).pack())
-                .build()
-        };
-        let witness_len: u64 = witness.as_bytes().len() as u64;
-        let mut hasher = new_blake2b();
-        hasher.update(tx_hash.as_bytes());
-        hasher.update(&witness_len.to_le_bytes());
-        hasher.update(&witness.as_bytes());
-        let message = {
-            let mut buf = [0u8; 32];
-            hasher.finalize(&mut buf);
-            H256::from(buf)
-        };
-        let sig = privkey.sign_recoverable(&message).expect("sign");
-        let witness = WitnessArgs::new_builder()
-            .lock(Some(Bytes::from(sig.serialize())).pack())
-            .build();
-        // sign input2
-        let witness2 = WitnessArgs::new_builder()
-            .lock(Some(Bytes::from(vec![0u8; 65])).pack())
-            .build();
-        let witness2_len: u64 = witness2.as_bytes().len() as u64;
-        let mut hasher = new_blake2b();
-        hasher.update(tx_hash.as_bytes());
-        hasher.update(&witness2_len.to_le_bytes());
-        hasher.update(&witness2.as_bytes());
-        let message2 = {
-            let mut buf = [0u8; 32];
-            hasher.finalize(&mut buf);
-            H256::from(buf)
-        };
-        let sig2 = privkey2.sign_recoverable(&message2).expect("sign");
-        let witness2 = WitnessArgs::new_builder()
-            .lock(Some(Bytes::from(sig2.serialize())).pack())
-            .build();
-        let tx = tx
-            .as_advanced_builder()
-            .witness(witness.as_bytes().pack())
-            .witness(witness2.as_bytes().pack())
-            .build();
-
-        let serialized_size = tx.data().as_slice().len() as u64;
-
-        assert_eq!(
-            serialized_size, TWO_IN_TWO_OUT_BYTES,
-            "2 in 2 out tx serialized size changed, PLEASE UPDATE consensus"
-        );
-
-        let (secp256k1_blake160_cell, secp256k1_blake160_cell_data) =
-            secp256k1_blake160_sighash_cell(consensus.clone());
-
-        let (secp256k1_data_cell, secp256k1_data_cell_data) = secp256k1_data_cell(consensus);
-
-        let input_cell1 = CellOutput::new_builder()
-            .capacity(capacity_bytes!(100).pack())
-            .lock(lock)
-            .build();
-
-        let resolved_input_cell1 =
-            CellMetaBuilder::from_cell_output(input_cell1, Default::default())
-                .out_point(input1.previous_output())
-                .build();
-
-        let input_cell2 = CellOutput::new_builder()
-            .capacity(capacity_bytes!(100).pack())
-            .lock(lock2)
-            .build();
-
-        let resolved_input_cell2 =
-            CellMetaBuilder::from_cell_output(input_cell2, Default::default())
-                .out_point(input2.previous_output())
-                .build();
-
-        let resolved_secp256k1_blake160_cell = CellMetaBuilder::from_cell_output(
-            secp256k1_blake160_cell,
-            secp256k1_blake160_cell_data,
-        )
-        .build();
-
-        let resolved_secp_data_cell =
-            CellMetaBuilder::from_cell_output(secp256k1_data_cell, secp256k1_data_cell_data)
-                .build();
-
-        let rtx = ResolvedTransaction {
-            transaction: tx,
-            resolved_cell_deps: vec![resolved_secp256k1_blake160_cell, resolved_secp_data_cell],
-            resolved_inputs: vec![resolved_input_cell1, resolved_input_cell2],
-            resolved_dep_groups: vec![],
-        };
-
-        let max_cycles = TWO_IN_TWO_OUT_CYCLES;
-        let verifier = TransactionScriptsVerifierWithEnv::new();
-        let result = verifier.verify(ScriptVersion::V0, &rtx, max_cycles);
-        assert!(result.is_ok());
-        let cycle = result.unwrap();
-        assert!(cycle <= TWO_IN_TWO_OUT_CYCLES);
-        assert!(cycle >= TWO_IN_TWO_OUT_CYCLES - CYCLE_BOUND);
-    }
-
-    #[test]
-    fn check_vm_version() {
-        let vm_version_cell_data = Bytes::from(
-            std::fs::read(Path::new(env!("CARGO_MANIFEST_DIR")).join("testdata/vm_version"))
-                .unwrap(),
-        );
-        let vm_version_cell = CellOutput::new_builder()
-            .capacity(Capacity::bytes(vm_version_cell_data.len()).unwrap().pack())
-            .build();
-        let vm_version_script = Script::new_builder()
-            .hash_type(ScriptHashType::Data1.into())
-            .code_hash(CellOutput::calc_data_hash(&vm_version_cell_data))
-            .build();
-        let output = CellOutputBuilder::default()
-            .capacity(capacity_bytes!(100).pack())
-            .lock(vm_version_script)
-            .build();
-        let input = CellInput::new(OutPoint::null(), 0);
-
-        let transaction = TransactionBuilder::default().input(input).build();
-
-        let dummy_cell = CellMetaBuilder::from_cell_output(output, Bytes::new())
-            .transaction_info(default_transaction_info())
-            .build();
-        let vm_version_cell =
-            CellMetaBuilder::from_cell_output(vm_version_cell, vm_version_cell_data)
-                .transaction_info(default_transaction_info())
-                .build();
-
-        let rtx = ResolvedTransaction {
-            transaction,
-            resolved_cell_deps: vec![vm_version_cell],
-            resolved_inputs: vec![dummy_cell],
-            resolved_dep_groups: vec![],
-        };
-
-        let verifier = TransactionScriptsVerifierWithEnv::new();
-        let result = verifier.verify_without_limit(ScriptVersion::V1, &rtx);
-        assert!(result.is_ok());
-    }
-
-    #[test]
-    fn check_exec_from_cell_data() {
-        let exec_caller_cell_data = Bytes::from(
-            std::fs::read(
-                Path::new(env!("CARGO_MANIFEST_DIR")).join("testdata/exec_caller_from_cell_data"),
-            )
-            .unwrap(),
-        );
-        let exec_caller_cell = CellOutput::new_builder()
-            .capacity(Capacity::bytes(exec_caller_cell_data.len()).unwrap().pack())
-            .build();
-
-        let exec_callee_cell_data = Bytes::from(
-            std::fs::read(Path::new(env!("CARGO_MANIFEST_DIR")).join("testdata/exec_callee"))
-                .unwrap(),
-        );
-        let exec_callee_cell = CellOutput::new_builder()
-            .capacity(Capacity::bytes(exec_callee_cell_data.len()).unwrap().pack())
-            .build();
-
-        let exec_caller_script = Script::new_builder()
-            .hash_type(ScriptHashType::Data1.into())
-            .code_hash(CellOutput::calc_data_hash(&exec_caller_cell_data))
-            .build();
-        let output = CellOutputBuilder::default()
-            .capacity(capacity_bytes!(100).pack())
-            .lock(exec_caller_script)
-            .build();
-        let input = CellInput::new(OutPoint::null(), 0);
-
-        let transaction = TransactionBuilder::default().input(input).build();
-
-        let dummy_cell = CellMetaBuilder::from_cell_output(output, Bytes::new())
-            .transaction_info(default_transaction_info())
-            .build();
-        let exec_caller_cell =
-            CellMetaBuilder::from_cell_output(exec_caller_cell, exec_caller_cell_data)
-                .transaction_info(default_transaction_info())
-                .build();
-
-        let exec_callee_cell =
-            CellMetaBuilder::from_cell_output(exec_callee_cell, exec_callee_cell_data)
-                .transaction_info(default_transaction_info())
-                .build();
-
-        let rtx = ResolvedTransaction {
-            transaction,
-            resolved_cell_deps: vec![exec_caller_cell, exec_callee_cell],
-            resolved_inputs: vec![dummy_cell],
-            resolved_dep_groups: vec![],
-        };
-
-        let verifier = TransactionScriptsVerifierWithEnv::new();
-        let result = verifier.verify_without_limit(ScriptVersion::V1, &rtx);
-        assert!(result.is_ok());
-    }
-
-    #[test]
-    fn check_exec_from_witness() {
-        let exec_caller_cell_data = Bytes::from(
-            std::fs::read(
-                Path::new(env!("CARGO_MANIFEST_DIR")).join("testdata/exec_caller_from_witness"),
-            )
-            .unwrap(),
-        );
-        let exec_caller_cell = CellOutput::new_builder()
-            .capacity(Capacity::bytes(exec_caller_cell_data.len()).unwrap().pack())
-            .build();
-
-        let exec_callee = Bytes::from(
-            std::fs::read(Path::new(env!("CARGO_MANIFEST_DIR")).join("testdata/exec_callee"))
-                .unwrap(),
-        )
-        .pack();
-
-        let exec_caller_script = Script::new_builder()
-            .hash_type(ScriptHashType::Data1.into())
-            .code_hash(CellOutput::calc_data_hash(&exec_caller_cell_data))
-            .build();
-        let output = CellOutputBuilder::default()
-            .capacity(capacity_bytes!(100).pack())
-            .lock(exec_caller_script)
-            .build();
-        let input = CellInput::new(OutPoint::null(), 0);
-
-        let transaction = TransactionBuilder::default()
-            .input(input)
-            .set_witnesses(vec![exec_callee])
-            .build();
-
-        let dummy_cell = CellMetaBuilder::from_cell_output(output, Bytes::new())
-            .transaction_info(default_transaction_info())
-            .build();
-        let exec_caller_cell =
-            CellMetaBuilder::from_cell_output(exec_caller_cell, exec_caller_cell_data)
-                .transaction_info(default_transaction_info())
-                .build();
-
-        let rtx = ResolvedTransaction {
-            transaction,
-            resolved_cell_deps: vec![exec_caller_cell],
-            resolved_inputs: vec![dummy_cell],
-            resolved_dep_groups: vec![],
-        };
-
-        let verifier = TransactionScriptsVerifierWithEnv::new();
-        let result = verifier.verify_without_limit(ScriptVersion::V1, &rtx);
-        assert!(result.is_ok());
-    }
-
-    #[test]
-    fn check_type_id_one_in_one_out_chunk() {
-        let (always_success_cell, always_success_cell_data, always_success_script) =
-            always_success_cell();
-        let always_success_out_point = OutPoint::new(h256!("0x11").pack(), 0);
-
-        let type_id_script = Script::new_builder()
-            .args(Bytes::from(h256!("0x1111").as_ref()).pack())
-            .code_hash(TYPE_ID_CODE_HASH.pack())
-            .hash_type(ScriptHashType::Type.into())
-            .build();
-
-        let input = CellInput::new(OutPoint::new(h256!("0x1234").pack(), 8), 0);
-        let input_cell = CellOutputBuilder::default()
-            .capacity(capacity_bytes!(1000).pack())
-            .lock(always_success_script.clone())
-            .type_(Some(type_id_script.clone()).pack())
-            .build();
-
-        let output_cell = CellOutputBuilder::default()
-            .capacity(capacity_bytes!(990).pack())
-            .lock(always_success_script.clone())
-            .type_(Some(type_id_script).pack())
-            .build();
-
-        let transaction = TransactionBuilder::default()
-            .input(input.clone())
-            .output(output_cell)
-            .cell_dep(
-                CellDep::new_builder()
-                    .out_point(always_success_out_point.clone())
-                    .build(),
-            )
-            .build();
-
-        let resolved_input_cell = CellMetaBuilder::from_cell_output(input_cell, Bytes::new())
-            .out_point(input.previous_output())
-            .build();
-        let resolved_always_success_cell = CellMetaBuilder::from_cell_output(
-            always_success_cell.clone(),
-            always_success_cell_data.to_owned(),
-        )
-        .out_point(always_success_out_point)
-        .build();
-
-        let rtx = ResolvedTransaction {
-            transaction,
-            resolved_cell_deps: vec![resolved_always_success_cell],
-            resolved_inputs: vec![resolved_input_cell],
-            resolved_dep_groups: vec![],
-        };
-
-        let mut cycles = 0;
-        let verifier = TransactionScriptsVerifierWithEnv::new();
-
-        verifier.verify_map(ScriptVersion::V0, &rtx, |verifier| {
-            let mut groups: Vec<_> = verifier.groups().collect();
-            let mut tmp: Option<ResumableMachine<'_>> = None;
-
-            loop {
-                if let Some(mut vm) = tmp.take() {
-                    cycles += vm.cycles();
-                    vm.set_cycles(0);
-                    match vm.machine.run() {
-                        Ok(code) => {
-                            if code == 0 {
-                                cycles += vm.cycles();
-                            } else {
-                                unreachable!()
-                            }
-                        }
-                        Err(error) => match error {
-                            VMInternalError::InvalidCycles => {
-                                tmp = Some(vm);
-                                continue;
-                            }
-                            _ => unreachable!(),
-                        },
-                    }
-                }
-                while let Some((ty, _, group)) = groups.pop() {
-                    let max = match ty {
-                        ScriptGroupType::Lock => ALWAYS_SUCCESS_SCRIPT_CYCLE - 10,
-                        ScriptGroupType::Type => TYPE_ID_CYCLES,
-                    };
-                    match verifier
-                        .verify_group_with_chunk(&group, max, &None)
-                        .unwrap()
-                    {
-                        ChunkState::Completed(used_cycles) => {
-                            cycles += used_cycles;
-                        }
-                        ChunkState::Suspended(vm) => {
-                            tmp = Some(vm);
-                            break;
-                        }
-                    }
-                }
-
-                if tmp.is_none() {
-                    break;
-                }
-            }
-        });
-
-        assert_eq!(cycles, TYPE_ID_CYCLES + ALWAYS_SUCCESS_SCRIPT_CYCLE);
-    }
-
-    #[test]
-    fn check_typical_secp256k1_blake160_2_in_2_out_tx_with_chunk() {
-        let consensus = ckb_testnet_consensus();
-        let dep_group_tx_hash = consensus.genesis_block().transactions()[1].hash();
-        let secp_out_point = OutPoint::new(dep_group_tx_hash, 0);
-
-        let cell_dep = CellDep::new_builder()
-            .out_point(secp_out_point)
-            .dep_type(DepType::DepGroup.into())
-            .build();
-
-        let input1 = CellInput::new(OutPoint::new(h256!("0x1234").pack(), 0), 0);
-        let input2 = CellInput::new(OutPoint::new(h256!("0x1111").pack(), 0), 0);
-
-        let mut generator = Generator::non_crypto_safe_prng(42);
-        let privkey = generator.gen_privkey();
-        let pubkey_data = privkey.pubkey().expect("Get pubkey failed").serialize();
-        let lock_arg = Bytes::from((&blake2b_256(&pubkey_data)[0..20]).to_owned());
-        let privkey2 = generator.gen_privkey();
-        let pubkey_data2 = privkey2.pubkey().expect("Get pubkey failed").serialize();
-        let lock_arg2 = Bytes::from((&blake2b_256(&pubkey_data2)[0..20]).to_owned());
-
-        let lock = Script::new_builder()
-            .args(lock_arg.pack())
-            .code_hash(type_lock_script_code_hash().pack())
-            .hash_type(ScriptHashType::Type.into())
-            .build();
-
-        let lock2 = Script::new_builder()
-            .args(lock_arg2.pack())
-            .code_hash(type_lock_script_code_hash().pack())
-            .hash_type(ScriptHashType::Type.into())
-            .build();
-
-        let output1 = CellOutput::new_builder()
-            .capacity(capacity_bytes!(100).pack())
-            .lock(lock.clone())
-            .build();
-        let output2 = CellOutput::new_builder()
-            .capacity(capacity_bytes!(100).pack())
-            .lock(lock2.clone())
-            .build();
-        let tx = TransactionBuilder::default()
-            .cell_dep(cell_dep)
-            .input(input1.clone())
-            .input(input2.clone())
-            .output(output1)
-            .output(output2)
-            .output_data(Default::default())
-            .output_data(Default::default())
-            .build();
-
-        let tx_hash: H256 = tx.hash().unpack();
-        // sign input1
-        let witness = {
-            WitnessArgs::new_builder()
-                .lock(Some(Bytes::from(vec![0u8; 65])).pack())
-                .build()
-        };
-        let witness_len: u64 = witness.as_bytes().len() as u64;
-        let mut hasher = new_blake2b();
-        hasher.update(tx_hash.as_bytes());
-        hasher.update(&witness_len.to_le_bytes());
-        hasher.update(&witness.as_bytes());
-        let message = {
-            let mut buf = [0u8; 32];
-            hasher.finalize(&mut buf);
-            H256::from(buf)
-        };
-        let sig = privkey.sign_recoverable(&message).expect("sign");
-        let witness = WitnessArgs::new_builder()
-            .lock(Some(Bytes::from(sig.serialize())).pack())
-            .build();
-        // sign input2
-        let witness2 = WitnessArgs::new_builder()
-            .lock(Some(Bytes::from(vec![0u8; 65])).pack())
-            .build();
-        let witness2_len: u64 = witness2.as_bytes().len() as u64;
-        let mut hasher = new_blake2b();
-        hasher.update(tx_hash.as_bytes());
-        hasher.update(&witness2_len.to_le_bytes());
-        hasher.update(&witness2.as_bytes());
-        let message2 = {
-            let mut buf = [0u8; 32];
-            hasher.finalize(&mut buf);
-            H256::from(buf)
-        };
-        let sig2 = privkey2.sign_recoverable(&message2).expect("sign");
-        let witness2 = WitnessArgs::new_builder()
-            .lock(Some(Bytes::from(sig2.serialize())).pack())
-            .build();
-        let tx = tx
-            .as_advanced_builder()
-            .witness(witness.as_bytes().pack())
-            .witness(witness2.as_bytes().pack())
-            .build();
-
-        let serialized_size = tx.data().as_slice().len() as u64;
-
-        assert_eq!(
-            serialized_size, TWO_IN_TWO_OUT_BYTES,
-            "2 in 2 out tx serialized size changed, PLEASE UPDATE consensus"
-        );
-
-        let (secp256k1_blake160_cell, secp256k1_blake160_cell_data) =
-            secp256k1_blake160_sighash_cell(consensus.clone());
-
-        let (secp256k1_data_cell, secp256k1_data_cell_data) = secp256k1_data_cell(consensus);
-
-        let input_cell1 = CellOutput::new_builder()
-            .capacity(capacity_bytes!(100).pack())
-            .lock(lock)
-            .build();
-
-        let resolved_input_cell1 =
-            CellMetaBuilder::from_cell_output(input_cell1, Default::default())
-                .out_point(input1.previous_output())
-                .build();
-
-        let input_cell2 = CellOutput::new_builder()
-            .capacity(capacity_bytes!(100).pack())
-            .lock(lock2)
-            .build();
-
-        let resolved_input_cell2 =
-            CellMetaBuilder::from_cell_output(input_cell2, Default::default())
-                .out_point(input2.previous_output())
-                .build();
-
-        let resolved_secp256k1_blake160_cell = CellMetaBuilder::from_cell_output(
-            secp256k1_blake160_cell,
-            secp256k1_blake160_cell_data,
-        )
-        .build();
-
-        let resolved_secp_data_cell =
-            CellMetaBuilder::from_cell_output(secp256k1_data_cell, secp256k1_data_cell_data)
-                .build();
-
-        let rtx = ResolvedTransaction {
-            transaction: tx,
-            resolved_cell_deps: vec![resolved_secp256k1_blake160_cell, resolved_secp_data_cell],
-            resolved_inputs: vec![resolved_input_cell1, resolved_input_cell2],
-            resolved_dep_groups: vec![],
-        };
-
-        let mut cycles = 0;
-        let verifier = TransactionScriptsVerifierWithEnv::new();
-        let result = verifier.verify_map(ScriptVersion::V0, &rtx, |verifier| {
-            let mut groups: Vec<_> = verifier.groups().collect();
-            let mut tmp: Option<ResumableMachine<'_>> = None;
-
-            loop {
-                if let Some(mut vm) = tmp.take() {
-                    cycles += vm.cycles();
-                    vm.set_cycles(0);
-                    match vm.machine.run() {
-                        Ok(code) => {
-                            if code == 0 {
-                                cycles += vm.cycles();
-                            } else {
-                                unreachable!()
-                            }
-                        }
-                        Err(error) => match error {
-                            VMInternalError::InvalidCycles => {
-                                tmp = Some(vm);
-                                continue;
-                            }
-                            _ => unreachable!(),
-                        },
-                    }
-                }
-                while let Some((_, _, group)) = groups.pop() {
-                    match verifier
-                        .verify_group_with_chunk(&group, TWO_IN_TWO_OUT_CYCLES / 10, &None)
-                        .unwrap()
-                    {
-                        ChunkState::Completed(used_cycles) => {
-                            cycles += used_cycles;
-                        }
-                        ChunkState::Suspended(vm) => {
-                            tmp = Some(vm);
-                            break;
-                        }
-                    }
-                }
-
-                if tmp.is_none() {
-                    break;
-                }
-            }
-
-            verifier.verify(TWO_IN_TWO_OUT_CYCLES)
-        });
-
-        let cycles_once = result.unwrap();
-        assert!(cycles <= TWO_IN_TWO_OUT_CYCLES);
-        assert!(cycles >= TWO_IN_TWO_OUT_CYCLES - CYCLE_BOUND);
-        assert_eq!(cycles, cycles_once);
-    }
 }
