+++ script/src/verify/tests/ckb_latest/mod.rs
@@ -1,0 +1,4 @@
+const SCRIPT_VERSION: crate::ScriptVersion = crate::ScriptVersion::latest();
+
+mod features_since_v2019;
+mod features_since_v2021;
