+++ script/src/verify/tests/mod.rs
@@ -1,0 +1,5 @@
+pub(crate) mod utils;
+
+mod ckb_2019;
+#[path = "ckb_latest/mod.rs"]
+mod ckb_2021;
